Implement server provisioning (#300)
<!-- Please make sure there is an issue that this PR is correlated to. -->
Depends on https://github.com/rivet-gg/rfcs/pull/7
Merge with https://github.com/rivet-gg/rivet-ee/pull/6
## Changes
- Move server provisioning to a service for future automation and flexibility
<!-- If there are frontend changes, please include screenshots. -->

+++ svc/pkg/upload/standalone/provider-fill/src/lib.rs
@@ -39,7 +39,7 @@ pub async fn run_from_env() -> GlobalResult<()> {
 		SET provider = $1
 		WHERE provider IS NULL 
 		",
-		proto_provider as i32 as i64,
+		proto_provider as i64,
 	)
 	.await?;
